Refactor start available test
