+++ tests/cli/workflow/mip/test_cli_mip_dna.py
@@ -145,6 +145,3 @@ def test_mip_dna_dev_start(
         use_bwa_mem=use_bwa_mem,
         start_with=start_with,
     )
-
-def test_mip_dna_dev_start_available():
-    #
