+++ tests/cli/workflow/nf_analysis/test_cli_start.py
@@ -11,11 +11,8 @@ from pytest_mock import MockerFixture
 from cg.apps.lims import LimsAPI
 from cg.cli.workflow.base import workflow as workflow_cli
 from cg.cli.workflow.raredisease.base import dev_start as raredisease_start
-from cg.cli.workflow.raredisease.base import dev_start_available as raredisease_start_available
 from cg.cli.workflow.rnafusion.base import start as rnafusion_start
-from cg.cli.workflow.rnafusion.base import start_available as rnafusion_start_available
 from cg.cli.workflow.taxprofiler.base import dev_start as taxprofiler_start
-from cg.cli.workflow.taxprofiler.base import dev_start_available as taxprofiler_start_available
 from cg.constants import EXIT_FAIL, EXIT_SUCCESS, Workflow
 from cg.meta.workflow.nf_analysis import NfAnalysisAPI
 from cg.meta.workflow.raredisease import RarediseaseAnalysisAPI
@@ -146,43 +143,3 @@ def test_start_nextflow_calls_service(
 
     # THEN the command should have executed without fail
     assert result.exit_code == EXIT_SUCCESS
-
-
-@pytest.mark.parametrize(
-    "start_available_command",
-    [raredisease_start_available, rnafusion_start_available, taxprofiler_start_available],
-    ids=["raredisease", "RNAFUSION", "Taxprofiler"],
-)
-@pytest.mark.parametrize(
-    "succeeds, exit_status",
-    [
-        (True, EXIT_SUCCESS),
-        (False, EXIT_FAIL),
-    ],
-)
-def test_start_available_nextflow_calls_service(
-    start_available_command: BaseCommand,
-    succeeds: bool,
-    exit_status: int,
-    cli_runner: CliRunner,
-    cg_context: CGConfig,
-    mocker: MockerFixture,
-):
-    """
-    Test that the start_available command succeeds when the starter succeeds and aborts otherwise.
-    """
-    # GIVEN a sufficient context
-
-    # GIVEN a mocked AnalysisStarter that simulates the start_available method
-    service_call: MagicMock = mocker.patch.object(
-        AnalysisStarter, "start_available", return_value=succeeds
-    )
-
-    # WHEN running the start-available command
-    result: Result = cli_runner.invoke(start_available_command, obj=cg_context)
-
-    # THEN the analysis starter should have been called with the expected input
-    service_call.assert_called_once()
-
-    # THEN the command should have executed without fail
-    assert result.exit_code == exit_status
